+++ Server/GPlayerServer/CHANGELOG.md
@@ -2,3 +2,7 @@
 
 Thu Feb 13 2025 14:34 - v0.0.2
   - refactor: 修改变量名，去除无效代码，暂时关闭jsoncpp使用
+  - refactor: 重新整理文件结构
+  - script: 初始化构建脚本auto_build.sh
+  - feat: 增加GXCodec转码器
+  - fix: 添加vscode查找路径，完善.vscode